restructure child nodes
