+++ packages/jaspr/lib/src/server/adapters/client_component_adapter.dart
@@ -1,4 +1,5 @@
 import '../../framework/framework.dart';
+import '../child_nodes.dart';
 import '../markup_render_object.dart';
 import '../server_binding.dart';
 import 'client_script_adapter.dart';
