+++ packages/jaspr/lib/src/server/adapters/element_boundary_adapter.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import '../../framework/framework.dart';
+import '../child_nodes.dart';
 import '../markup_render_object.dart';
 import '../server_binding.dart';
