+++ packages/jaspr/lib/src/server/document/document.dart
@@ -7,6 +7,7 @@ import 'package:html/parser.dart';
 
 import '../../../server.dart';
 import '../adapters/element_boundary_adapter.dart';
+import '../child_nodes.dart';
 import '../markup_render_object.dart';
 
 export '../../components/style.dart' hide Style;
